more realistic time left evaluation
Evaluate the time left only when a percentage point of the total battery
level changes.

Otherwise we end up to have very bad fluctuations in the time left
estimation: the measure of the battery level in percentage points is too
coarse grained, so we end up to have an estimation more close the real
value when the battery level changes and a too much optimistic when the
battery is actually discharging but the percentage point is not yet
changed.

Signed-off-by: Andrea Righi <andrea@betterlinux.com>